Extract UrlGenerator to make adapter better testable
