+++ src/phpDocumentor/Transformer/Writer/Twig/LinkRenderer/LinkAdapter.php
@@ -14,26 +14,16 @@ declare(strict_types=1);
 namespace phpDocumentor\Transformer\Writer\Twig\LinkRenderer;
 
 use InvalidArgumentException;
-use OutOfRangeException;
-use phpDocumentor\Descriptor\Descriptor;
 use phpDocumentor\Descriptor\DescriptorAbstract;
-use phpDocumentor\Descriptor\GuideSetDescriptor;
 use phpDocumentor\Path;
 use phpDocumentor\Reflection\DocBlock\Tags\Reference;
 use phpDocumentor\Reflection\Fqsen;
 use phpDocumentor\Reflection\Type;
 use phpDocumentor\Reflection\Types\Object_;
-use phpDocumentor\Transformer\Router\Router;
 use phpDocumentor\Transformer\Writer\Twig\LinkRenderer;
 use phpDocumentor\Transformer\Writer\Twig\LinkRendererInterface;
-use Webmozart\Assert\Assert;
 
 use function is_string;
-use function ltrim;
-use function sprintf;
-use function strlen;
-use function strpos;
-use function substr;
 
 /**
  * Renders most of the links to elements, urls and urls with virtual schemes.
@@ -46,14 +36,14 @@ use function substr;
 final class LinkAdapter implements LinkRendererInterface
 {
     private LinkRenderer $rendererChain;
-    private Router $router;
+    private UrlGenerator $urlGenerator;
     private HtmlFormatter $formatter;
 
-    public function __construct(LinkRenderer $rendererChain, Router $router, HtmlFormatter $formatter)
+    public function __construct(LinkRenderer $rendererChain, UrlGenerator $urlGenerator, HtmlFormatter $formatter)
     {
         $this->rendererChain = $rendererChain;
-        $this->router = $router;
         $this->formatter = $formatter;
+        $this->urlGenerator = $urlGenerator;
     }
 
     /**
@@ -73,7 +63,7 @@ final class LinkAdapter implements LinkRendererInterface
 
         $target = new Target(
             $this->determineTitle($resolvedValue),
-            $this->generateUrl($resolvedValue, (string) $value),
+            $this->urlGenerator->generate($resolvedValue, (string) $value),
             $this->normalizePresentation($resolvedValue, $presentation)
         );
 
@@ -113,104 +103,6 @@ final class LinkAdapter implements LinkRendererInterface
         }
 
         return $target;
-    }
-
-    /**
-     * @param string|Path|Type|DescriptorAbstract|Fqsen|Reference\Reference|Reference\Fqsen $target
-     */
-    private function generateUrl($target, string $fallback): ?string
-    {
-        $unlinkable = $target instanceof Fqsen || $target instanceof Type;
-        if ($unlinkable) {
-            // With an unlinkable object, we don't know if the page for it exists; so we don't render a link to it.
-            return null;
-        }
-
-        if ($this->isGuideUrl($target)) {
-            return $this->generateGuideUrl($target);
-        }
-
-        if (!$target instanceof Descriptor) {
-            return $fallback;
-        }
-
-        try {
-            $url = $this->router->generate($target);
-        } catch (InvalidArgumentException $e) {
-            return null;
-        }
-
-        return $this->withoutLeadingSlash($url);
-    }
-
-    private function withoutLeadingSlash(string $path): string
-    {
-        return ltrim($path, '/');
-    }
-
-    /**
-     * @param string|Reference\Url $target
-     */
-    private function generateGuideUrl($target): ?string
-    {
-        if ((is_string($target) || $target instanceof Reference\Url) === false) {
-            throw new InvalidArgumentException(
-                'Guide references can only be derived from a string or Url reference'
-            );
-        }
-
-        $guideReference = substr((string) $target, strlen('doc://'));
-
-        // TODO: of course this is not correct, but phpDocumentor does not have a mechanism yet for determining the
-        // current version and is prepared for multi-version support but doesn't actually have it yet
-        $currentVersion = $this->rendererChain->getProject()->getVersions()->first();
-        if ($currentVersion === null) {
-            // unlinkable, thus no URL
-            return null;
-        }
-
-        // TODO: of course this is not correct, but phpDocumentor does not have a mechanism yet for determining the
-        // right guide set and it is prepared for multi-guide support but doesn't actually have it yet
-        /** @var ?GuideSetDescriptor $guideSet */
-        $guideSet = $currentVersion->getDocumentationSets()->filter(GuideSetDescriptor::class)->first();
-        if ($guideSet === null) {
-            // unlinkable, thus no URL
-            return null;
-        }
-
-        Assert::isInstanceOf($guideSet, GuideSetDescriptor::class);
-
-        // TODO: This is copied from the TableOfContentsBuilder; once we get proper support for Versions,
-        // we need to revisit this.
-        $documentEntry = $guideSet->getMetas()->findDocument($guideReference);
-        if ($documentEntry === null) {
-            return null;
-        }
-
-        try {
-            $document = $guideSet->getDocuments()->get($documentEntry->getFile());
-        } catch (OutOfRangeException $e) {
-            return null;
-        }
-
-        return sprintf(
-            '%s/%s',
-            $guideSet->getOutputLocation(),
-            // TODO: Add Support for DocumentEntries to the router
-            $this->withoutLeadingSlash($this->router->generate($document))
-        );
-    }
-
-    /**
-     * @param string|Reference\Url $target
-     */
-    private function isGuideUrl($target): bool
-    {
-        if ($target instanceof Reference\Url) {
-            $target = (string) $target;
-        }
-
-        return is_string($target) && strpos($target, 'doc://') === 0;
     }
 
     /**
